requirements.txt
praw
python-dotenv
pyyaml
google-api-python-client